Change unint8 to int for UI and fix e_stop topic

+++ mission_planner/include/mission_planner/EStopButton.h
@@ -8,18 +8,18 @@ class EStopButton : public QPushButton {
     Q_OBJECT
 public:
     EStopButton(QWidget* qwidget) : QPushButton(qwidget) { }
-    void setType(const bool isEStop, const uint8_t robotNum);
+    void setType(const bool isEStop, const int robotNum);
     ~EStopButton();
 
 signals:
-    void clicked(bool isEStop, uint8_t robotNum);
+    void clicked(bool isEStop, int robotNum);
 
 private slots:
     void reemitClicked();
 
 private:
     bool isEStop;
-    uint8_t robotNum;
+    int robotNum;
 };
